+++ src/pathselection/BeaconLinkQualityMeasurement.cpp
@@ -245,7 +245,7 @@ BeaconLinkQualityMeasurement::BeaconLinkQualityMeasurement(wns::ldk::fun::FUN* f
     for (int ii = 0; ii<config.len("localIDs.keys()"); ++ii)
     {
         std::string key = config.get<std::string>("localIDs.keys()",ii);
-        uint32_t value  = config.get<uint32_t>("localIDs.values()",ii);
+        unsigned long int value  = config.get<unsigned long int>("localIDs.values()",ii);
         registry.addProvider(wns::probe::bus::contextprovider::Constant(key, value));
         MESSAGE_SINGLE(VERBOSE, logger, "Using Local IDName '"<<key<<"' with value: "<<value);
     }
